+++ src/main/kotlin/nl/deschepers/laraveltinker/linemarkerprovider/TinkerRunLineMarkerProvider.kt
@@ -4,8 +4,10 @@ import com.intellij.codeInsight.daemon.LineMarkerInfo
 import com.intellij.codeInsight.daemon.LineMarkerProvider
 import com.intellij.openapi.editor.markup.GutterIconRenderer.Alignment.RIGHT
 import com.intellij.openapi.util.IconLoader
+import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.elementType
+import com.jetbrains.php.lang.psi.stubs.PhpFileElementType
 import nl.deschepers.laraveltinker.Strings
 import nl.deschepers.laraveltinker.types.LaravelTinkerConsoleType
 import nl.deschepers.laraveltinker.repository.ConsoleFileRepository
@@ -17,15 +19,18 @@ class TinkerRunLineMarkerProvider : LineMarkerProvider {
     }
 
     override fun getLineMarkerInfo(element: PsiElement): LineMarkerInfo<*>? {
-        val consoleFileRepository = ConsoleFileRepository(element.project)
-        if (element.containingFile.virtualFile.fileType is LaravelTinkerConsoleType && element.elementType.toString() == PHP_OPEN_TAG) {
+        val isTinkerConsole = element.containingFile.virtualFile.fileType is LaravelTinkerConsoleType
+        val isOpenTag = element.elementType.toString() == PHP_OPEN_TAG
+        val isEndOfFile = element.textRange.endOffset == element.containingFile.textRange.endOffset && element.elementType is PhpFileElementType
+
+        if (isTinkerConsole && (isOpenTag || isEndOfFile)) {
             return LineMarkerInfo(
                 element,
-                element.textRange,
+                TextRange(element.textRange.endOffset, element.textRange.endOffset),
                 IconLoader.getIcon("icons/tinker-run.svg", javaClass),
                 null,
                 { _: MouseEvent, psiElement: PsiElement ->
-                    consoleFileRepository.runFile(psiElement.containingFile.virtualFile)
+                    ConsoleFileRepository(element.project).runFile(psiElement.containingFile.virtualFile)
                 },
                 RIGHT,
                 { Strings.get("lt.run") }