+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Run line marker now also show up at the last line of the tinker console.
+
 ## [3.0.0] - 2025-06-19
 
 ### Added